SPEC CHANGE: instantiate defaults to model class instead of CouchTiny::Document
This means that an expression like
    Foo.get "id"
    Foo.view_by_bar ...
which gets a document without a type attribute, or one which is unknown,
will create an instance of Foo instead of CouchTiny::Document

This can be overridden by a new default_class parameter to the low-level
CouchTiny::Document.instantiate method (but is not yet passed through as an
option from finders)

If the document does have a type attribute, this still takes precedence, e.g.
   Foo.get "id"
will create an instance of Bar if the document has {"type":"Bar"} and you
have created a Bar model class.